refactor(value): collapse the five copy-pasted .cache LazyList arms
ADR-0038 phase 4 (deliberately deferred cleanup, no behaviour change).
ADR-0038 S1.7/S1.8 found that the `.cache` LazyList arm
(`ll.is_genuinely_lazy() || ll.is_cat_pull()` ->
`with_cached_no_sink().with_list_context()`) was copy-pasted at five
call sites, and that the three `__mutsu_lazylist_*` markers behind it
were stringly-keyed entries in the value's captured closure `env`
rather than typed LazyList fields.

- Add three typed bool fields to LazyList (array_context, list_context,
  cached_no_sink), parallel to the SeqView field ADR-0038 added to
  SeqBody. The existing accessor/mutator methods keep their names and
  signatures, so no caller outside value_lazy.rs needed to change.
- Add LazyList::cache_lazy_view(), a single helper implementing the
  .cache arm once; all five call sites (builtins/methods_0arg,
  runtime/methods_call_dispatch, and the three VM call-method paths)
  now call it instead of duplicating the guard + tagged-clone logic.
- Split the new_* constructors and the scan-reduction forcer out of
  value_lazy.rs into a new value_lazy_ctors.rs, since the file had
  already drifted past the repo's 500-line convention and this change
  pushed it further; both files are now under 500 lines.

Verified with t/seq-cache-returns-list.t (the ADR-0038 oracle-agreement
pin), the full t/seq-*.t / t/lazy-seq-*.t / t/*lazy*.t /
t/io-cathandle-lazy.t suites, and a before/after comparison of the
MUTSU_REAL_TEST=1 roast files ADR-0038 phase 3 fixed confirming
identical behaviour.
