+++ src/builtins/functions/dispatch_1arg.rs
@@ -156,6 +156,9 @@ pub(crate) fn native_function_1arg(name: &str, arg: &Value) -> Option<Result<Val
                             closure_seq: None,
                             walk_pending: None,
                             cat_pull: None,
+                            array_context: false,
+                            list_context: false,
+                            cached_no_sink: false,
                         };
                         return Some(Ok(Value::lazy_list(crate::gc::Gc::new(ll))));
                     }
@@ -183,6 +186,9 @@ pub(crate) fn native_function_1arg(name: &str, arg: &Value) -> Option<Result<Val
                     closure_seq: None,
                     walk_pending: None,
                     cat_pull: None,
+                    array_context: false,
+                    list_context: false,
+                    cached_no_sink: false,
                 };
                 return Some(Ok(Value::lazy_list(crate::gc::Gc::new(ll))));
             }
