+++ src/runtime/methods_dispatch_match2.rs
@@ -627,6 +627,9 @@ impl Interpreter {
             closure_seq: None,
             walk_pending: None,
             cat_pull: None,
+            array_context: false,
+            list_context: false,
+            cached_no_sink: false,
         };
         Value::lazy_list(crate::gc::Gc::new(list))
     }
