+++ src/vm/vm_call_method_mut_ops.rs
@@ -787,12 +787,10 @@ impl Interpreter {
         // the matching note in the non-mut dispatch path.
         if let ValueView::LazyList(ll) = target.view()
             && method == "cache"
-            && (ll.is_genuinely_lazy() || ll.is_cat_pull())
+            && let Some(result) = ll.cache_lazy_view()
         {
             crate::vm::vm_stats::record_dispatch_entry_intercept("callmethodmut", "lazy-cache");
-            self.stack.push(Value::lazy_list(crate::gc::Gc::new(
-                ll.with_cached_no_sink().with_list_context(),
-            )));
+            self.stack.push(result);
             return Ok(());
         }
         let target = if let ValueView::LazyList(ll) = target.view()
