+++ src/vm/vm_call_method_ops.rs
@@ -1141,12 +1141,10 @@ impl Interpreter {
         // mid-iteration changes to the cat's `.nl-in`/`.chomp`.
         if let ValueView::LazyList(ll) = target.view()
             && method == "cache"
-            && (ll.is_genuinely_lazy() || ll.is_cat_pull())
+            && let Some(result) = ll.cache_lazy_view()
         {
             crate::vm::vm_stats::record_dispatch_entry_intercept("callmethod", "lazy-cache");
-            self.stack.push(Value::lazy_list(crate::gc::Gc::new(
-                ll.with_cached_no_sink().with_list_context(),
-            )));
+            self.stack.push(result);
             return Ok(());
         }
         let target = if let ValueView::LazyList(ll) = target.view()
